+++ pkg/api-server/catalog_ws_test.go
@@ -15,6 +15,8 @@ var _ = Describe("Catalog WS", func() {
 	It("should return the api catalog", func() {
 		// given
 		cfg := config.DefaultApiServerConfig()
+		cfg.Catalog.Admin.LocalUrl = "http://localhost:1111"
+		cfg.Catalog.Admin.PublicUrl = "https://kuma.internal:2222"
 		cfg.Catalog.DataplaneToken.LocalUrl = "http://localhost:1111"
 		cfg.Catalog.DataplaneToken.PublicUrl = "https://kuma.internal:2222"
 		cfg.Catalog.Bootstrap.Url = "http://kuma.internal:3333"
@@ -51,6 +53,10 @@ var _ = Describe("Catalog WS", func() {
 					"url": "http://kuma.internal:3333"
 				},
 				"dataplaneToken": {
+					"localUrl": "http://localhost:1111",
+					"publicUrl": "https://kuma.internal:2222"
+				},
+				"admin": {
 					"localUrl": "http://localhost:1111",
 					"publicUrl": "https://kuma.internal:2222"
 				}