+++ pkg/config/api-server/catalog/config.go
@@ -3,7 +3,8 @@ package catalog
 // Not yet exposed via YAML and env vars on purpose
 type CatalogConfig struct {
 	Bootstrap      BootstrapApiConfig
-	DataplaneToken DataplaneTokenApiConfig
+	DataplaneToken DataplaneTokenApiConfig // DEPRECATED: remove in next major version of Kuma
+	Admin          AdminApiConfig
 }
 
 type BootstrapApiConfig struct {
@@ -11,6 +12,11 @@ type BootstrapApiConfig struct {
 }
 
 type DataplaneTokenApiConfig struct {
+	LocalUrl  string
+	PublicUrl string
+}
+
+type AdminApiConfig struct {
 	LocalUrl  string
 	PublicUrl string
 }